+++ uxarray/__init__.py
@@ -1,5 +1,5 @@
 # from uxarray.utils.helpers import *
-from uxarray.utils.constants import (INT_DTYPE, INT_FILL_VALUE)
+from uxarray.constants import (INT_DTYPE, INT_FILL_VALUE)
 from uxarray.grid.grid import Grid
 from uxarray.core.api import (open_grid, open_dataset, open_mfdataset)
 from uxarray.core.dataarray import UxDataArray
