+++ uxarray/grid/connectivity.py
@@ -1,7 +1,7 @@
 import numpy as np
 import xarray as xr
 
-from uxarray.utils.constants import INT_DTYPE, INT_FILL_VALUE
+from uxarray.constants import INT_DTYPE, INT_FILL_VALUE
 
 
 def close_face_nodes(Mesh2_face_nodes, nMesh2_face, nMaxMesh2_face_nodes):
