+++ uxarray/io/_mpas.py
@@ -2,7 +2,7 @@ import xarray as xr
 import numpy as np
 import warnings
 
-from uxarray.utils.constants import INT_DTYPE, INT_FILL_VALUE
+from uxarray.constants import INT_DTYPE, INT_FILL_VALUE
 
 
 def _primal_to_ugrid(in_ds, out_ds):
