--- uxarray/utils/constants.py
@@ -1,5 +1,0 @@
-import numpy as np
-
-# numpy indexing code is written for np.intp
-INT_DTYPE = np.intp
-INT_FILL_VALUE = np.iinfo(INT_DTYPE).min
